Shift main to cmd dir
